pyproject.toml
[tool.poetry]
name = "airflow-tui"
version = "0.1.0"
description = "A Textual User Interface (TUI) built with Textual for Apache Airflow"
authors = ["Tomás Farías Santana <tomas@tomasfarias.dev>"]
license = "MIT"
readme = "README.md"

classifiers = [
    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.8"
textual = "^0.1.15"

[tool.poetry.dev-dependencies]
apache-airflow = "^2.2.3"
black = "^22.1"
flake8 = "^4.0"
isort = "^5.10"
mypy = "^0.931"
pytest = "^6.2"
pre-commit = "^2.17"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
